requirements.txt
Keras==2.4.3
opencv-python==4.4.0.44
numpy==1.18.5
Django==3.1.1
djangorestframework==3.12.1
tensorflow==2.3.1